fixed conflicts and imports due to repo restructure
