+++ go.mod
@@ -5,13 +5,17 @@ go 1.21.0
 require github.com/golang/protobuf v1.5.3
 
 require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/net v0.15.0 // indirect
 	golang.org/x/sys v0.12.0 // indirect
 	golang.org/x/text v0.13.0 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20230920204549-e6e6cdab5c13 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
 require (
+	github.com/stretchr/testify v1.8.4
 	google.golang.org/grpc v1.58.2
 	google.golang.org/protobuf v1.31.0
 )
