+++ internal/protocol/utils/protobuf_formatting.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	pb "github.com/Domilz/d7017e-mesh-network/internal/protocol/grpc_bidirectional_stream/protofiles/tag"
+	pb "github.com/Domilz/d7017e-mesh-network/internal/protocol/protofiles/tag"
 )
 
 // PrintFormattedState prints a formatted State message with multiple Readings.
